plutus-tutorial: update for TH removal
Mostly removing things. Also made the indentation consistently 4 spaces.
